+++ gem/main.py
@@ -32,6 +32,7 @@ from glob import glob
 from shutil import copy2 as copy
 
 # System
+from sys import exit as sys_exit
 from argparse import ArgumentParser
 
 # ------------------------------------------------------------------------------